% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVDJ.R
\name{addVDJ}
\alias{addVDJ}
\title{Add 10x VDJ Data to Seurat Obj}
\usage{
addVDJ(seurat_obj, type_v, dir_v, barcodeString_v = "1", flag_v = T)
}
\arguments{
\item{seurat_obj}{gene expression seurat object}

\item{type_v}{either "t" for TCR or "b" for BCR}

\item{dir_v}{directory where "filtered_contig_annotations.csv" and "clonotypes.csv" are found.}

\item{barcodeString_v}{see details. CHARACTER string indicating what cell barcodes to grab.
Default is 1, other values used when loading from multiple samples)}

\item{flag_v}{default is True, which will add a column "hasT" or "hasB" (depends on type_v) that will have TRUE if that cell has
TCR/BCR info and FALSE if it does not.}
}
\value{
Exact same seurat object with new metadata columns with clonotype information:
\link{type_v}_clonotype_id and \link{type_v}_cdr3s_aa
}
\description{
Add 10x VDJ T or B data to an initialized GEX Seurat object. Modified from function found:
https://ucdavis-bioinformatics-training.github.io/2020-Advanced_Single_Cell_RNA_Seq/data_analysis/VDJ_Analysis_fixed
}
\details{
Cell barcodes are of the format \link{ATCG}+\-1 in an individual seurat object. If cellranger aggregate is run on the TCR/BCR output,
then it solves the problem of barcode duplication by changing the "1" at the end with another number. In the case I used to work on this (batch3),
the replaced values are the sample numbers - I'm not sure if this is because of arguments provided in aggregate,
or if it just goes through numerically and does it. Regardless, if you're unsure take a peak at the barcode suffixes in contigs_dt
}
