% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atacMarkerGenes_v}
\alias{atacMarkerGenes_v}
\title{ATAC Marker Genes}
\format{
\subsection{\code{atacMarkerGenes_v}}{

Vector of length 21
\describe{
\item{values}{gene names}
}
}
}
\usage{
atacMarkerGenes_v
}
\description{
Vector containing genes used for ATAC major population designations
}
\keyword{datasets}
