% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b2_lymphoidColors2_v}
\alias{atlas_b2_lymphoidColors2_v}
\title{Batch2 Lymphoid Colors 2}
\format{
\subsection{\code{atlas_b2_lymphoidColors2_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
atlas_b2_lymphoidColors2_v
}
\description{
Colors used on UMAPs for b2 lymphoid populations
second option.
}
\keyword{datasets}
