% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b2_myeloidColors2_v}
\alias{atlas_b2_myeloidColors2_v}
\title{Batch2 Myeloid Colors 2}
\format{
\subsection{\code{atlas_b2_myeloidColors2_v}}{

Named vector of length 7
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
atlas_b2_myeloidColors2_v
}
\description{
Colors used on UMAPs for b2 myeloid populations
when collapsing like sub-pops.
}
\keyword{datasets}
