% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b3_lymphoidColors_v}
\alias{atlas_b3_lymphoidColors_v}
\title{Batch3 Lymphoid Colors}
\format{
\subsection{\code{atlas_b3_lymphoidColors_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
atlas_b3_lymphoidColors_v
}
\description{
Colors used on UMAPs for b3 lymphoid populations
}
\keyword{datasets}
