% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b3_myeloidColors_v}
\alias{atlas_b3_myeloidColors_v}
\title{Batch3 Myeloid Colors}
\format{
\subsection{\code{atlas_b3_myeloidColors_v}}{

Named vector of length 6
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
atlas_b3_myeloidColors_v
}
\description{
Colors used on UMAPs for b3 myeloid populations
}
\keyword{datasets}
