% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b3_neoplasticColors_v}
\alias{atlas_b3_neoplasticColors_v}
\title{Batch3 Neoplastic Colors}
\format{
\subsection{\code{atlas_b3_neoplasticColors_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name (neo.c#)}
}
}
}
\usage{
atlas_b3_neoplasticColors_v
}
\description{
Colors used on UMAPs for b3 neplastic populations
}
\keyword{datasets}
