% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlas_b3_otherColors_v}
\alias{atlas_b3_otherColors_v}
\title{Batch3 Other Colors}
\format{
\subsection{\code{atlas_b3_otherColors_v}}{

Named vector of length 3
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
atlas_b3_otherColors_v
}
\description{
Colors used on UMAPs for b3 other populations
}
\keyword{datasets}
