% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atlasData.R
\docType{data}
\name{atlasGeneLists_lsv}
\alias{atlasGeneLists_lsv}
\title{Atlas Gene Lists}
\format{
\subsection{\code{atlasGeneLists_lsv}}{

Named list of length 3
\describe{
\item{values}{vector of gene names}
\itme{names}{name of gene list}
}
}
}
\usage{
atlasGeneLists_lsv
}
\description{
Gene lists used for atlas project
\enumerate{
\item 15 MMTvPyMT genes
\item 6 'neoplastic' genes
\item 1 'myeloid' gene
}
}
\keyword{datasets}
