% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgConservedMarkers.R
\name{avgConservedMarkers}
\alias{avgConservedMarkers}
\title{Average Conserved Markers}
\usage{
avgConservedMarkers(
  conserved_dt,
  geneCol_v = "Gene",
  extraCols_v = c("gene_biotype", "description")
)
}
\arguments{
\item{conserved_dt}{output of FindConservedMarkers, converted to data.table.}

\item{geneCol_v}{column name of gene names}

\item{extraCols_v}{any columns that aren't in default FCM() output. Default here is gene_biotype which is added when I merge output with annotations.}
}
\value{
data.table with same number of rows as conserved_dt, but only 'Gene', 'max_pval', 'minimump_p_val', 'avg_padj' (calculated in script), and 'avg_log2FC' (calculated in script)
}
\description{
take the output of FindConservedMarkers and average the log2FC and adj p-value of all the comparisons
}
