% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_bCellColors_v}
\alias{b12_bCellColors_v}
\title{Batch1+2 B cell Sublcuster colors}
\format{
\subsection{\code{b12_bCellColors_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{B Cell sub population}
}
}
}
\usage{
b12_bCellColors_v
}
\description{
Colors used on UMAPs for b1+b2 b cell populations
}
\keyword{datasets}
