% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_bCellPops_dt}
\alias{b12_bCellPops_dt}
\title{Batch1+2 B Cell Pops}
\format{
\subsection{\code{b12_bCellPops_dt}}{

data.table with 5 rows and 2 columns
\describe{
\item{cluster}{seurat cluster assignment}
\item{pop}{population label}
}
}
}
\usage{
b12_bCellPops_dt
}
\description{
Data.table that matches seurat clusters with
B cell sub-populations
}
\keyword{datasets}
