% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_lymphoidBadColors_v}
\alias{b12_lymphoidBadColors_v}
\title{Batch1+2 Lymphoid Colors 2}
\format{
\subsection{\code{b12_lymphoidBadColors_v}}{

Named vector of length 11
\describe{
\item{values}{Hex Color}
\item{names}{Collapsed Sub Population Name}
}
}
}
\usage{
b12_lymphoidBadColors_v
}
\description{
Colors used on UMAPs for b1+b2 lymphoid populations
before sub-pops are collapsed.
}
\keyword{datasets}
