% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_lymphoidColors_v}
\alias{b12_lymphoidColors_v}
\title{Batch1+2 Lymphoid Colors}
\format{
\subsection{\code{b12_lymphoidColors_v}}{

Named vector of length 6
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
b12_lymphoidColors_v
}
\description{
Colors used on UMAPs for b1+b2 lymphoid populations
}
\keyword{datasets}
