% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_myeloidBadColors_v}
\alias{b12_myeloidBadColors_v}
\title{Batch1+2 Myeloid Colors 2}
\format{
\subsection{\code{b12_myeloidBadColors_v}}{

Named vector of length 7
\describe{
\item{values}{Hex Color}
\item{names}{Collapsed Sub Population Name}
}
}
}
\usage{
b12_myeloidBadColors_v
}
\description{
Colors used on UMAPs for b1+b2 myeloid populations
before sub-pops are collapsed.
}
\keyword{datasets}
