% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_myeloidColors_v}
\alias{b12_myeloidColors_v}
\title{Batch1+2 Myeloid Colors}
\format{
\subsection{\code{b12_myeloidColors_v}}{

Named vector of length 7
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
b12_myeloidColors_v
}
\description{
Colors used on UMAPs for b1+b2 myeloid populations
}
\keyword{datasets}
