% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_myeloidPopMap_dt}
\alias{b12_myeloidPopMap_dt}
\title{Batch1+2 Myeloid Pop Map}
\format{
\subsection{\code{b12_lymphoidPopMap_dt}}{

data.table with 7 rows and 2 columns
\describe{
\item{sPop}{original sPop labels}
\item{collapsePop}{collapsed population labels}
}
}
}
\usage{
b12_myeloidPopMap_dt
}
\description{
Data.table that matches original sPops
with collapsed populations
}
\keyword{datasets}
