% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_neoplasticColors_v}
\alias{b12_neoplasticColors_v}
\title{Batch12 Neoplastic Colors}
\format{
\subsection{\code{b12_neoplasticColors_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name (neo.c#)}
}
}
}
\usage{
b12_neoplasticColors_v
}
\description{
Colors used on UMAPs for b12 neplastic populations
}
\keyword{datasets}
