% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_res_v}
\alias{b12_res_v}
\title{Batch1+2 Resolution values}
\format{
\subsection{\code{b12_res_v}}{

Named vector of length 4
\describe{
\item{values}{resolution}
\item{names}{seurat object name}
}
}
}
\usage{
b12_res_v
}
\description{
Resolution values used for clustering in batch12 seurat object
}
\keyword{datasets}
