% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b12_treats_v}
\alias{b12_treats_v}
\title{Batch1+2 Treatment Labels}
\format{
\subsection{\code{b12_treats_v}}{

vector of length 7
\describe{
\item{values}{"4x", "3xR", "3xNR", "2x", "Ent", "PTX", "Unt"}
}
}
}
\usage{
b12_treats_v
}
\description{
Treatment names
}
\keyword{datasets}
