% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_clusterNames_v}
\alias{b3_clusterNames_v}
\title{Batch3 Cluster names}
\format{
\subsection{\code{b3_clusterNames_v}}{

Named vector of length 4
\describe{
\item{values}{meta.data column name}
\item{names}{seurat object name}
}
}

\subsection{\code{b3_umapNames_v}}{

Named vector of length 4
\describe{
\item{values}{umap name from seurat object}
\item{names}{seurat object name}
}
}

\subsection{\code{b3_res_v}}{

Named vector of length 4
\describe{
\item{values}{resolution}
\item{names}{seurat object name}
}
}
}
\usage{
b3_clusterNames_v

b3_clusterNames_v

b3_clusterNames_v
}
\description{
Column names found in batch3 seurat object
metadata that contain the clusters identified
from the Seurat analysis pipeline. The names
are formated "seurat_clusters_d\link{dim}_res\link{res}
where dim = dimensions used in reduction and
res = resolution used for clustering

UMAP names found in names() of batch3 seurat object

Resolution values used for clustering in batch3 seurat object
}
\keyword{datasets}
