% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_lymphoid2Colors_v}
\alias{b3_lymphoid2Colors_v}
\title{Batch3 Lymphoid Colors 2}
\format{
\subsection{\code{b3_lymphoid2Colors_v}}{

Named vector of length 5
\describe{
\item{values}{Hex Color}
\item{names}{Collapsed Sub Population Name}
}
}
}
\usage{
b3_lymphoid2Colors_v
}
\description{
Colors used on UMAPs for b3 lymphoid populations
when collapsing like sub-pops.
}
\keyword{datasets}
