% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_lymphoidFullColors_v}
\alias{b3_lymphoidFullColors_v}
\title{Batch3 Lymphoid Full Colors}
\format{
\subsection{\code{b3_lymphoidFullColors_v}}{

Named vector of length 12
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
b3_lymphoidFullColors_v
}
\description{
Colors used on UMAPs for b3 lymphoid populations (with Neoplastic)
}
\keyword{datasets}
