% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_myeloidBadColors_v}
\alias{b3_myeloidBadColors_v}
\title{Batch3 Myeloid Colors 2}
\format{
\subsection{\code{b3_myeloidBadColors_v}}{

Named vector of length 11
\describe{
\item{values}{Hex Color}
\item{names}{Collapsed Sub Population Name}
}
}
}
\usage{
b3_myeloidBadColors_v
}
\description{
Colors used on UMAPs for b3 myeloid populations
before sub-pops collapsed.
}
\keyword{datasets}
