% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_myeloidColors_v}
\alias{b3_myeloidColors_v}
\title{Batch3 Myeloid Colors}
\format{
\subsection{\code{b3_myeloidColors_v}}{

Named vector of length 11
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
b3_myeloidColors_v
}
\description{
Colors used on UMAPs for b3 myeloid populations
}
\keyword{datasets}
