% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_myeloidPopMap_dt}
\alias{b3_myeloidPopMap_dt}
\title{Batch3 Myeloid Pop Map}
\format{
\subsection{\code{b3_lymphoidPopMap_dt}}{

data.table with 11 rows and 2 columns
\describe{
\item{sPop}{original sPop labels}
\item{collapsePop}{collapsed population labels}
}
}
}
\usage{
b3_myeloidPopMap_dt
}
\description{
Data.table that matches original sPops
with collapsed populations
}
\keyword{datasets}
