% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_neoplasticColors_v}
\alias{b3_neoplasticColors_v}
\title{Batch3 Neoplastic Colors}
\format{
\subsection{\code{b3_neoplasticColors_v}}{

Named vector of length 13
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name (neo.c#)}
}
}
}
\usage{
b3_neoplasticColors_v
}
\description{
Colors used on UMAPs for b3 neplastic populations
}
\keyword{datasets}
