% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_treats_v}
\alias{b3_treats_v}
\title{Batch3 Treatment Labels}
\format{
\subsection{\code{b3_treats_v}}{

vector of length 8
\describe{
\item{values}{"4x", "3xR", "3xNR", "2x", "Ent", "PTX", "NT", "d80"}
}
}
}
\usage{
b3_treats_v
}
\description{
Treatment names
}
\keyword{datasets}
