% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b3_umapNames_v}
\alias{b3_umapNames_v}
\title{Batch3 UMAP names}
\format{
\subsection{\code{b3_umapNames_v}}{

Named vector of length 4
\describe{
\item{values}{umap name from seurat object}
\item{names}{seurat object name}
}
}
}
\usage{
b3_umapNames_v
}
\description{
UMAP names found in names() of batch3 seurat object
}
\keyword{datasets}
