% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmbeddings.R
\name{calcEmbeddings}
\alias{calcEmbeddings}
\title{Calculate Feature Embeddings}
\usage{
calcEmbeddings(seurat_obj, clusterName_v, reductionName_v, breaks_v = T, ...)
}
\arguments{
\item{seurat_obj}{seurat object with data reduction}

\item{clusterName_v}{name of meta data column containing cluster identities to calculate embeddings for}

\item{reductionName_v}{name of the reduction that embeddings are calculated from}

\item{breaks_v}{logical. Should color breaks be adjusted or not}

\item{...}{extra parameters to pass to determineBreaks}
}
\value{
data.frame of embeddings and prints a heatmap
}
\description{
Calculate feature embeddings for a given single cell reduction
}
