% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcViolinStatPositions.R
\name{calcViolinStatPositions}
\alias{calcViolinStatPositions}
\title{Calculate Violin Stat Positions}
\usage{
calcViolinStatPositions(
  plot_df,
  measureVar_v,
  cm_df,
  levels_v,
  yOffset_v = 1.56
)
}
\arguments{
\item{plot_df}{data.frame used for plotting}

\item{measureVar_v}{name of column in plot_df that contains metric to be plotted}

\item{cm_df}{data.frame of compare_means output}

\item{levels_v}{factor levels. needed to get x-position}

\item{yOffset_v}{amount to multiply maximum y-value by to set y-limit}
}
\value{
named list of length 2:
\enumerate{
\item "cm_df" contains cm_df with the added xPos and yPos columns
\item "ylim" contains y limits for the plot
}
}
\description{
determine xy coordinates for stat::compare_means results
}
