% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGenesInData.R
\name{checkGenesInData}
\alias{checkGenesInData}
\title{Check Genes In Data}
\usage{
checkGenesInData(
  obj,
  objName_v,
  geneList_v,
  geneListName_v = "",
  verbose_v = T
)
}
\arguments{
\item{obj}{seurat object to check}

\item{objName_v}{name of seurat object for update}

\item{geneList_v}{vector of gene names to check. OR named list of vectors of gene names}

\item{geneListName_v}{name of gene list for output. Only used if geneList_v is a vector.}

\item{verbose_v}{logical. true - print updates. false - no printing}
}
\value{
same gene list, but subset for only those that are in obj.
}
\description{
See if provided genes are present in the dataset
}
