% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNames.R
\name{checkNames}
\alias{checkNames}
\title{Check Object and DEG Names}
\usage{
checkNames(obj, deg)
}
\arguments{
\item{obj}{seurat object}

\item{deg}{deg table}
}
\description{
Make sure that seurat object and DEG table have same pop names
}
