% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSCTResCol.R
\name{cleanSCTResCol}
\alias{cleanSCTResCol}
\title{Clean SCT Res Col}
\usage{
cleanSCTResCol(obj, grep_v = "SCT_snn.*", which_v = 2)
}
\arguments{
\item{obj}{seurat object with multiple clustering resolutions}

\item{grep_v}{regex used to grab the columns to be cleaned.}

\item{which_v}{which of the results returned by grep_v should be kept?}
}
\value{
same input seurat object, but only one resolution column
}
\description{
remove extra resolution columns in obj@meta.data
}
\details{
When processing the data, there are two runs of SCT with and without doublets.
We want the second one, without doublets and here will remove the first. This fxn can be
used with other parameter values, but the defaults are what are needed for this project.
}
