% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myClusterSweep.R
\name{clusterQC}
\alias{clusterQC}
\title{Cluster QC}
\usage{
clusterQC(seurat_obj, embedding_v, ndims_v, reductionName_v, reduction_v)
}
\arguments{
\item{seurat_obj}{A seurat object with dimensional reduction embeddings calculated (PCA)}

\item{embedding_v}{embeddings for chosen dimensional reduction. Usually: seruat_obj@reductions$pca@cell.embeddings}

\item{ndims_v}{number of dimensions to use. Default is 10.}

\item{reductionName_v}{name of reduction to use.}

\item{reduction_v}{reduction to use}
}
\value{
List containing the following elements:
"QC" - tibble of mean silhouette and RMSD values for each resolution
"dimPlot" - list of DimPlots, one for each resolution
"silBox" - list of silhouette boxplots, one for each resolution
"RMSDBar" - list of RMSD barplots, one for each resolution
}
\description{
Given a seurat object and embeddings, calculate qc
}
