% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDADE.R
\name{compareDADE}
\alias{compareDADE}
\title{Compare Vs Each DEGs with ATAC}
\usage{
compareDADE(
  atacDAG_ls,
  baseDEG_ls,
  pop_v,
  treat1_v = "4x",
  treat2_v = "3xR",
  lfc_v = 0.5,
  pval_v = 0.05,
  immunoAntigens_v = NULL,
  allAntigens_v = NULL
)
}
\arguments{
\item{atacDAG_ls}{list of DAG results (element name format is paste0(treat1_v, "\emph{vs}", treat2_v))}

\item{baseDEG_ls}{list of DEG results (e.g. vsEach$batch3_neoplastic or global)}

\item{pop_v}{name of population to grab from baseDEG_ls (e.g. c8 for vsEach$batch3_neoplastic$c8, or batch3_neoplastic for global$batch3_neoplastic)}

\item{treat1_v}{name of first treatment from comparison}

\item{treat2_v}{name of second treatment from comparison}

\item{lfc_v}{log fold change cut-off}

\item{pval_v}{p value cut-off}

\item{immunoAntigens_v}{vector of immunogenic antigens}

\item{allAntigens_v}{vector of all antigens}

\item{atacGenes_lsv}{list of atac DA results (one up and one down)}
}
\value{
tbd
}
\description{
Compare Vs Each DEGs with ATAC
}
\details{
atacGenes_lsv must be list in format list("Up" = someVector, "Down" = someVector)
}
