% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareMarkerResults.R
\name{compareMarkerResults}
\alias{compareMarkerResults}
\title{Compare Marker Results Log-Fold-Change}
\usage{
compareMarkerResults(
  data1_dt,
  data2_dt,
  name_v = NULL,
  ident1_v,
  ident2_v,
  identCol_v,
  meta_dt = NULL,
  geneCol_v = "Gene",
  runNames_v = c("sct", "rna"),
  lfc_v = 0.5,
  pval_v = 0.05,
  toPlot_v = c("scatter", "volcano"),
  compareName_v,
  verbose_v = F
)
}
\arguments{
\item{data1_dt}{findMarker results}

\item{data2_dt}{findmarker results}

\item{name_v}{character of global name of what is presented. Usually lymphoid, neoplastic, etc.}

\item{ident1_v}{first ident in findmarker call}

\item{ident2_v}{second ident in findMarker call}

\item{identCol_v}{column name of meta.data that corresponds to idents}

\item{meta_dt}{optional meta.data. If not provided, will take cell counts from objects.}

\item{geneCol_v}{column name in data1_dt and data2_dt that correspond to genes. Default is "Gene"}

\item{runNames_v}{vector of length two specifying the names for data1_dt and data2_dt.}

\item{lfc_v}{log fold change to use for cut-offs}

\item{pval_v}{adjusted p-value to use for cut-offs}

\item{toPlot_v}{vector of plot types to output. Current options are 'scatter' and 'volcano'}

\item{compareName_v}{passed to splitVar_v in plotVolcano. Name to describe what's different between data1_dt/data2_dt (i.e. why we're comparing them)}

\item{verbose_v}{passed to plotVolcano. Print messages.}
}
\description{
plot lfc results of different findMarkers runs
}
\details{
Take two different FindMarkers() results and compare the log-fold change outputs.
Written to compare runs between different assays (SCT vs RNA), but could compare other things, I think.
}
