% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePCs.R
\name{comparePCs}
\alias{comparePCs}
\title{Compare top n genes of PCs from different seurat objects}
\usage{
comparePCs(
  obj1,
  obj2,
  dr1 = "pca",
  dr2 = NULL,
  dim1,
  dim2 = NULL,
  nFeat_v = 100,
  which_v = "all"
)
}
\arguments{
\item{obj1}{First seurat object}

\item{obj2}{Second seurat object}

\item{dr1}{Name of dimensional reduction to grab for obj1. Default is 'pca'}

\item{dr2}{Name of dimensional reduction to grab fro obj2. If NULL (default) will be same as dr1.}

\item{dim1}{Index of principal component to use for obj1}

\item{dim2}{Index of principal component to use for obj2. If NULL (default) will be same as dim1.}

\item{nFeat_v}{Number of features to compare.}

\item{which_v}{Which genes to compare. "all" takes both positive and negative genes;
"positive" is positive only; "negative" is negative only}
}
\value{
Vector of gene names shared between the two PCs
}
\description{
Compare the top genes of a given PC from two different seurat objects
}
