% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareVsEachWithATAC.R
\name{compareVsEachWithATAC}
\alias{compareVsEachWithATAC}
\title{Compare Vs Each DEGs with ATAC}
\usage{
compareVsEachWithATAC(
  atacGenes_lsv,
  baseDEG_ls,
  pop_v,
  treat1_v = "4x",
  treat2_v = "3xR",
  lfc_v = 0.5,
  pval_v = 0.05,
  immunoAntigens_v,
  allAntigens_v
)
}
\arguments{
\item{atacGenes_lsv}{list of atac DA results (one up and one down)}

\item{baseDEG_ls}{list of DEG results (e.g. vsEach$batch3_neoplastic)}

\item{pop_v}{name of population to grab from baseDEG_ls (e.g. c8 for vsEach$batch3_neoplastic$c8)}

\item{treat1_v}{name of first treatment from comparison}

\item{treat2_v}{name of second treatment from comparison}

\item{lfc_v}{log fold change cut-off}

\item{pval_v}{p value cut-off}

\item{immunoAntigens_v}{vector of immunogenic antigens}

\item{allAntigens_v}{vector of all antigens}
}
\value{
tbd
}
\description{
Compare Vs Each DEGs with ATAC
}
\details{
atacGenes_lsv must be list in format list("Up" = someVector, "Down" = someVector)
}
