% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaySummary.R
\name{displaySummary}
\alias{displaySummary}
\title{Display Summary Info}
\usage{
displaySummary(
  obj,
  subObj = NULL,
  name_v = NULL,
  subCol_v = "sPop",
  batch_v,
  cPops_v = F,
  popCol_v = "mPop",
  groupBy_v = list("Population", "Treatment", "Cluster"),
  alpha_v = 0.5,
  summary_lsv = list(treat = "Treatment", pop = "mPop", combo = c("Treatment", "mPop")),
  outDir_v = NULL,
  outName_v = NULL,
  print_v = T,
  displayOrder_v = "decreasing",
  cutOff_v = cutOff_v
)
}
\arguments{
\item{obj}{seurat object to summarize}

\item{subObj}{subset object. need to update getSummaryTables to work on subObj, so i don't need both of these}

\item{name_v}{population name. will remove "batch#", if present. See details.}

\item{subCol_v}{Column to use in conjunction with name_v to subset data to summarize.}

\item{batch_v}{name of batch}

\item{cPops_v}{logical indicating if the sub-populations are the "collapsed" pops or not. Used to get appropriate colors.}

\item{popCol_v}{column name for population annotations}

\item{groupBy_v}{vector containing any or all of 'Population', 'Treatment', and 'Cluster', indicating which UMAPs to make.}

\item{alpha_v}{alpha parameter only used for by-treatment plot.}

\item{summary_lsv}{list of vectors of treatments to summarize. See details.}

\item{outDir_v}{optional output directory to save plots and data}

\item{outName_v}{optional name for output file.}

\item{print_v}{logical indicating to print tables and plots to console.}

\item{displayOrder_v}{either 'decreasing' to indicate that identity with most number of cells goes on the bottom, or 'factor' to go by factor levels.}

\item{cutOff_v}{numeric value indicating minimum number of cells needed in each group.}
}
\value{
returns list of summary tables as well as two optional outputs:
\enumerate{
\item save plots and data to file
and/or
\item print plots to console.
}
}
\description{
Display summary tables and umaps of provided object.
}
\details{
This function calls getSummaryTables(), formats those tables, and makes umaps
of seurat clusters and population identities.

summary_lsv determines how the data are summarized. It is a named list where
each element's name is user defined and should describe what the output is summarizing.
Refer to the defaults for help. The values of each element correspond to the column or
columns that will be subset from obj@meta.data in order to generate that summary.

The default will summarize Treatment alone, major population alone, and treat x pop
together. Right now can only handle 'treat', 'pop' and 'combo'
}
