% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dotMarkers_v}
\alias{dotMarkers_v}
\title{Dot Plot Markers}
\format{
\subsection{\code{dotMarkers_v}}{

Named vector of length 97
\describe{
\item{values}{gene names}
\item{names}{population that gene corresponds to}
}
}
}
\usage{
dotMarkers_v
}
\description{
Vector containing unique genes from markers_lsv for dotplot output
}
\keyword{datasets}
