% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpCols_v}
\alias{fpCols_v}
\title{Feature Plot Color Scale}
\format{
\subsection{\code{fpCols_v}}{

vector of hex codes of length 9
\describe{
\item{value}{Hex Color}
}
}
}
\usage{
fpCols_v
}
\description{
Color scale to use in Feature plots
blue - green - yellow - orange - red
low - med - high
Provided by Sushma Nagaraj
}
\keyword{datasets}
