% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fullAbbrTreatmentMap_dt}
\alias{fullAbbrTreatmentMap_dt}
\title{scRNA/scATAC Treatment Code Mapping}
\format{
\subsection{\code{fullAbbrTreatmentMap_dt}}{

data.table with 22 rows and 2 columns
\describe{
\item{Treatment}{scRNAseq Treatment Code}
\item{longTreatment}{long-form treatment code used in scATAC}
\item{Name}{Full Rx Name}
}
}
}
\usage{
fullAbbrTreatmentMap_dt
}
\description{
data.table that maps shorthand treatment
codes to long-form "_"-separated codes. Note
that PTX_ENT_aCSF1R_unt is in sc
}
\keyword{datasets}
