% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeatGenes.R
\name{getHeatGenes}
\alias{getHeatGenes}
\title{Get Heatmap Genes - NOT UPDATED!!}
\usage{
getHeatGenes(
  data_dt,
  kind_v,
  geneCol_v = "Gene",
  lfcCol_v = "avg_log2FC",
  lfc_v = 0.25,
  pvalCol_v = "p_val_adj",
  pval_v = 0.05,
  labelAll_v = T
)
}
\arguments{
\item{data_dt}{data.table output from FindAllMarkers}

\item{kind_v}{either "sig" to grab significant genes or an integer to grab top # genes.}

\item{geneCol_v}{name of column containing gene names. Default is Gene}

\item{lfcCol_v}{name of column containing log fold change values. Default is avg_log2FC (used to subset top genes, optionally used to determine sig genes)}

\item{lfc_v}{log fold change cut-off for significance}

\item{pvalCol_v}{name of column containing adjusted pvalues. Default is p_val_adj (used to subset top and sig genes)}

\item{pval_v}{pvalue cut-off to use}

\item{labelAll_v}{passed to volcanoWrangleMarkers. TRUE - all sig genes are labeled. FALSE - only top 25 sig genes labeled}

\item{lfcLabelCutOff_v}{passed to volcanoWrangleMarkers. TRUE - use lfc_v threshold to determine significant genes in addition to pval_v}
}
\value{
list with two vectors. First, 'heat' are the genes to be displayed in the heatmap. Second, 'label' contains the genes that are significant.
If kind_v == "sig", then these will be the same. If kind_v = #, then it will be the significant genes and is used to add a * to their name in heatmap
}
\description{
filter find marker results to get genes for heatmap plotting
}
\details{
Need to update the label options to mirror volcanowrangleMarkers. Essentially need to add labelGenes_v argument.
}
