% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectRPlottingData.R
\name{getProjectRPlottingData}
\alias{getProjectRPlottingData}
\title{Get ProjectR Plotting Data}
\usage{
getProjectRPlottingData(
  obj,
  batch_v,
  pop_v,
  sigName_v = "sncScore",
  binaryName_v = "SnC",
  popCol_v = "collapsePop",
  metaCol_v = "Treatment"
)
}
\arguments{
\item{obj}{seurat object with new columns created by prepProjectRPlotting}

\item{batch_v}{name of batch (required to get appropriate treatments)}

\item{pop_v}{name of population (required to get appropriate population order)}

\item{sigName_v}{must be same as what was used in prepProjectRPlotting}

\item{binaryName_v}{must be same as what was used in prepProjectRPlotting}

\item{popCol_v}{name of meta.data columbn containing population information. Can usually be 'mPop', 'sPop', or 'collapsedPop'}

\item{metaCol_v}{name of the meta.data column containing grouping variable. Can be more than one, Treatment is required.}
}
\value{
data.frame
}
\description{
extract meta.data from seurat object for easier plotting with ggplot
}
