% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryTables.R
\name{getSummaryTables}
\alias{getSummaryTables}
\title{Get Tables}
\usage{
getSummaryTables(
  obj,
  subset_v = NULL,
  subCol_v = "mPop",
  summary_lsv = list(treat = "Treatment", pop = "mPop", combo = c("Treatment", "mPop"))
)
}
\arguments{
\item{obj}{seurat object to summarize. Must have specified columns}

\item{subset_v}{subset data on provided value}

\item{subCol_v}{specify which column to subset using subset_v}

\item{summary_lsv}{list of vectors of columns to summarize.}
}
\description{
Get summary tables of selected columns

Currently make 3 different tables: treatment, population, and treatment x population
Treatment should almost always be 'treat = Treatment'; population should be 'mPop' if subset_v = NULL and
'sPop' if not (or sPop2 or another version); treatment x population must agree with the others.
}
