% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreatmentComparisons.R
\name{getTreatmentComparisons}
\alias{getTreatmentComparisons}
\title{Get Treatment Comparisons}
\usage{
getTreatmentComparisons(treats_v)
}
\arguments{
\item{treats_v}{vector of treatments to run comparisons for}
}
\value{
list. One element per "treat1" comparison. Each "treat1" element has a vector
containing each "treat2" to be compared against it.
}
\description{
Create a list of pairwise treatment comparisons from an input vector of treatments

Returns a non-repeating list of pairs, so order of treats_v matters.
}
