% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gseaColors_v}
\alias{gseaColors_v}
\title{GSEA Plot Color Scale}
\format{
\subsection{\code{gseaColors_v}}{

named vector of color names of length 2
\describe{
\item{values}{red, blue}
\item{names}{UP, DOWN}
}
}
}
\usage{
gseaColors_v
}
\description{
Color scale to use in GSEA plots
blue - down
red - up
}
\keyword{datasets}
