% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelSigHeatGenes.R
\name{labelSigHeatGenes}
\alias{labelSigHeatGenes}
\title{Label genes for heatmap}
\usage{
labelSigHeatGenes(seurat_obj, genes_lsv, assay_v = "fullSCT")
}
\arguments{
\item{seurat_obj}{seurat object subset for the cells to plot in heatmap}

\item{genes_lsv}{list of 2 vectors, 'heat' are the genes to label, 'label' are those to add * to (only done if heat != label)}

\item{assay_v}{assay to use. Should usually be fullSCT}
}
\value{
seurat object with modified labels for genes and also vector of modified genes for doheatmap
}
\description{
Add a * to significant genes for heatmap output
}
