% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCustomGeneSets.R
\name{loadCustomGeneSets}
\alias{loadCustomGeneSets}
\title{Load Custom Gene Sets}
\usage{
loadCustomGeneSets(file_v, keep_v = NULL, omit_v = NULL)
}
\arguments{
\item{file_v}{path to excel file containing one or more sheets with gene lists. See details.}

\item{keep_v}{vector of sheet names to keep}

\item{omit_v}{vector of sheet names to omit.}
}
\value{
list of tables
}
\description{
Load custom gene sets
}
\details{
Only one of keep_v or omit_v may be set.
Gene list format is:
\enumerate{
\item one sheet per gene set
\item sheet name is gene set name
\item Must at least contain a column GENE with gene names.
}
}
