% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(
  markerDir_v = NULL,
  markerClass_v = "global",
  metaDir_v = NULL,
  rdsDir_v = NULL,
  batch_v,
  pop_v,
  toRead_v = "_2x_|_3xNR_|_3xR_|_4x_|_d80_|_Ent_|_NT_|_PTX_|_3x_|_3xR\\\\+4x|_Unt_"
)
}
\arguments{
\item{markerDir_v}{path to where marker results are held}

\item{markerClass_v}{either 'global' or 'vsEach', indicating which read in function to use.}

\item{metaDir_v}{path to metadata directory}

\item{rdsDir_v}{path to seurat objects}

\item{batch_v}{batch name that identifies files to read in}

\item{pop_v}{population names that identify files to read in.}

\item{toRead_v}{used to grep what to input.}
}
\value{
list with marker results, metadata, and seurat objects. Read in objects with provided directories. If dir == NULL, not read in.
}
\description{
Load DEG, meta, and seurat data
}
