% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMsigdbr.R
\name{loadMsigdbr}
\alias{loadMsigdbr}
\title{Load msigdbr}
\usage{
loadMsigdbr(
  species_v = "mmu",
  category_v = c("H", "C2", "C2", "C2", "C5", "C5"),
  subcategory_v = c("", "CP:KEGG_LEGACY", "CP:BIOCARTA", "CP:REACTOME", "GO:BP", ""),
  name_v = c("HALLMARK", "KEGG", "BIOCARTA", "REACTOME", "GOBP", "GO")
)
}
\arguments{
\item{species_v}{what species to use.
For mouse, use "Mouse", "mmu", "Mus musculus". For human, use "Human", "hg", "Homo sapiens"}

\item{category_v}{vector of categories to grab data for}

\item{subcategory_v}{vector of subcategories. Must be same length as category_v. Use "" if no subcat}

\item{name_v}{character vector of names to label outputs.}
}
\value{
list of lists of msigdbr results
}
\description{
Load various reference sets from msigdbr database
}
