% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lymphoidMinorPopGenes_v}
\alias{lymphoidMinorPopGenes_v}
\title{Lymphoid Minor Population Genes}
\format{
\subsection{\code{lymphoidMinorPopGenes_v}}{

Vector of length 14
\describe{
\item{values}{gene names}
}
}
}
\usage{
lymphoidMinorPopGenes_v
}
\description{
Vector containing genes used for lymphoid minor population designations
}
\keyword{datasets}
