% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{majorPopGenes_v}
\alias{majorPopGenes_v}
\title{Major Population Genes}
\format{
\subsection{\code{majorPopGenes_v}}{

Vector of length 23
\describe{
\item{values}{gene names}
}
}
}
\usage{
majorPopGenes_v
}
\description{
Vector containing genes used for major population designations
}
\keyword{datasets}
