% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDotPlots.R
\name{makeDotPlots}
\alias{makeDotPlots}
\title{Make Dot Plots}
\usage{
makeDotPlots(
  seurat_obj,
  name_v,
  assay_v = "fullSCT",
  features_v,
  cols_v = "RdYlBu",
  groupBy_v,
  splitBy_v = NULL
)
}
\arguments{
\item{seurat_obj}{object to use for plot}

\item{name_v}{name used for title}

\item{assay_v}{assay used for plot}

\item{features_v}{what genes to plot}

\item{cols_v}{dotplot colors}

\item{groupBy_v}{fed into the group.by argument, usually cluster or treatment}

\item{splitBy_v}{optional split.}
}
\value{
ggplot object
}
\description{
standard seurat dotplot wrapper
}
