% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRunTable.R
\name{makeRunTable}
\alias{makeRunTable}
\title{Make run Table}
\usage{
makeRunTable(list_ls, treats_v, ...)
}
\arguments{
\item{list_ls}{input list}

\item{treats_v}{treatment vector, used for ordering}

\item{...}{passed to orderTreatments}
}
\value{
data.table with one row per unique combination of columns
}
\description{
Use names from input list to determine what should be run.
}
