% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markerCols_lsv}
\alias{markerCols_lsv}
\title{Population Marker List Columns}
\format{
\subsection{\code{markerCols_lsv}}{

Named list of length 11
\describe{
\item{values}{number of columns to use}
\item{names}{name of gene lists}
}
}
}
\usage{
markerCols_lsv
}
\description{
List containing number of columns for feature plot output for each marker set.
}
\keyword{datasets}
