% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_lsv}
\alias{markers_lsv}
\title{Population Marker List}
\format{
\subsection{\code{markers_lsv}}{

Named list of length 11
\describe{
\item{values}{vector of gene names}
\item{names}{name of gene lists}
}
}
}
\usage{
markers_lsv
}
\description{
List of markers used to identify sub-populations
\enumerate{
\item tumor - 9 genes
\item Caf - 14 genes
\item MyoEpi - 17 genes
\item Endo - 9 genes
\item Mono - 5 genes
\item DC - 7 genes
\item TAMs - 9 genes
\item MDSC - 9 genes
\item Tcells - 10 genes
\item Bcells - 4 genes
\item NK - 4 genes
}
}
\keyword{datasets}
