% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mPopColors_v}
\alias{mPopColors_v}
\title{Major Population Colors}
\format{
\subsection{\code{mPopColors_v}}{

Named vector of length 6
\describe{
\item{values}{Hex Color}
\item{names}{Population Name}
}
}
}
\usage{
mPopColors_v
}
\description{
Colors used on UMAPs for major populations
}
\keyword{datasets}
