% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutationCounter.R
\name{mutationCounter}
\alias{mutationCounter}
\title{Mutation Counter}
\usage{
mutationCounter(
  mutation_dt,
  idCols_v = c("Hugo_Symbol", "Tumor_Sample_Barcode"),
  specificSummary_lsv = list(`4x` = c("I4", "I5", "I6")),
  specificCol_v = "Tumor_Sample_Barcode"
)
}
\arguments{
\item{mutation_dt}{WES mutation table}

\item{idCols_v}{column names of mutation_dt to use for mutation summary tables. Should generally be a gene ID and a sample ID}

\item{specificSummary_lsv}{list of vectors of sample IDs to use for a specific mutation count}

\item{specificCol_v}{column name of mutation_dt that is used for specific summary}
}
\value{
summary data.table with above columns, sort by decreasing nSamp
}
\description{
summarize WES mutation table in multiple ways
}
\details{
Summary columns
'nSamp' simple count of how many samples gene is mutated in
'nSampSpecific' simple count of how many specific samples gene is mutated in
'total' total count of all mutations in all samples of this gene
}
