% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{myeloidMinorPopGenes_v}
\alias{myeloidMinorPopGenes_v}
\title{Myeloid Minor Population Genes}
\format{
\subsection{\code{myeloidMinorPopGenes_v}}{

Vector of length 19
\describe{
\item{values}{gene names}
}
}
}
\usage{
myeloidMinorPopGenes_v
}
\description{
Vector containing genes used for myeloid minor population designations
}
\keyword{datasets}
