% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myTreatFacetUmap.R
\name{myTreatFacetUmap}
\alias{myTreatFacetUmap}
\title{My Treatment Facet Umap}
\usage{
myTreatFacetUmap(
  obj,
  umap_v,
  vals_v,
  colors_v,
  title_v = NULL,
  size_v = 36,
  facetVar_v = "Treatment",
  groupVar_v = "sPop",
  plotType_v = "umap"
)
}
\arguments{
\item{obj}{seurat object}

\item{umap_v}{reduction to use}

\item{vals_v}{the unique values of facetVar_v (need to be in order)}

\item{colors_v}{colors to plot with}

\item{title_v}{optional plot title}

\item{size_v}{size for plot title}

\item{facetVar_v}{variable to facet by (should be 'Treatment')}

\item{plotType_v}{either 'umap' or 'scatter'}
}
\description{
custom facet umap for figures
}
