% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neoantigenPlots.R
\name{neoantigenPlots}
\alias{neoantigenPlots}
\title{Neoantigen Plots}
\usage{
neoantigenPlots(
  obj,
  deg_dt,
  out_lsls,
  cells_v,
  ident1_v = "neo.c8",
  ident2_v,
  cellCounts_v
)
}
\arguments{
\item{obj}{corresponding seurat object}

\item{deg_dt}{Find markers output fed into wrangleDEG}

\item{out_lsls}{output of wrangleDEG}

\item{cells_v}{character vector indicating cells included in comparison (e.g. all or 4x)}

\item{ident1_v}{name of first identity in comparison. Used for titles and cell counts}

\item{ident2_v}{name of second identity. Used for titles and cell counts.}

\item{cellCounts_v}{vector of length 2 that indicates number of cells in each identity. Must be in order ident1, ident2}
}
\value{
list containing one dot plot and one volcano plot for each entry
}
\description{
Dot and Volcano for specific genes
}
