% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newCircos.R
\name{newCircos}
\alias{newCircos}
\title{Plot expression of a receptor's ligands by other cell types as a chord plot}
\usage{
newCircos(
  signaling_dt,
  dest,
  arc_genes,
  expression_threshold = 0.01,
  cellCol_v = "Pop",
  cell_idents = NULL,
  cell_colors = NULL,
  title = "Signaling",
  label_arcs = T,
  multi_plot = NULL,
  destExpr_v,
  max_v = NULL,
  plotHeight_v = 7
)
}
\arguments{
\item{signaling_dt}{Data.table of signaling expression. Must contain columns
for cell type (cellCol_v), 'origin', 'destination', and 'mean.expression'.}

\item{dest}{name of destination gene}

\item{arc_genes}{name(s) of ligands to include as arcs}

\item{expression_threshold}{Minimum mean expression value of a ligand by a cell type for a chord to be rendered between the cell type and the receptor}

\item{cellCol_v}{column name for cell types - usually either 'Pop' or 'Treatment'}

\item{cell_idents}{Vector of cell types from cluster assignments in the domino object to be included in the plot.}

\item{cell_colors}{Named vector of color names or hex codes where names correspond to the plotted cell types and the color values}

\item{title}{if provided, final title will be paste0(receptor/ligand, " Signaling")}

\item{label_arcs}{logical indicating if arc populations should be labeled}

\item{multi_plot}{numeric vector that, If provided, indicates position of this
plot in a multi-panel layout for legend placement. Even = right side,
odd = left side.}

\item{destExpr_v}{expression controlling arc width for destination gene}

\item{max_v}{optional global maximum. If NULL, uses local maximum}

\item{plotHeight_v}{height in inches of plot. Used to make the scale legend appropriately sized}
}
\value{
renders a circos plot to the active graphics device
}
\description{
Creates a chord plot of expression of ligands that can activate a specified
receptor where chord widths correspond to mean ligand expression by the cluster.
}
\details{
Sender arc widths are always set to 1.0 (unit width) and colored by ligand.
The destination arc width is set to 2 / nrow(signaling_df) so it remains
narrow relative to the sender arcs. Expression values are scaled to \link{0, 1}
relative to max_v (or local max) before being used as link widths.

The lgd_chord legend bar height is proportional to the sender arc width
relative to the total arc space: 1 / (nrow(signaling_df) + 1).
}
