% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oneGroupVolcanoColors_v}
\alias{oneGroupVolcanoColors_v}
\title{Volcano Color Scale - One group}
\format{
\subsection{\code{oneGroupVolcanoColors_v}}{

named vector of colors of length 3
\describe{
\item{values}{grey, blue, red}
\item{names}(NO, DOWN, UP)
}
}
}
\usage{
oneGroupVolcanoColors_v
}
\description{
Color scale to use in standard
volcano plots
red - higher expression
blue - lower expression
grey - no change in expression
}
\keyword{datasets}
