% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderTreatments.R
\name{orderTreatments}
\alias{orderTreatments}
\title{Order treatments}
\usage{
orderTreatments(
  names_lsv,
  order_v,
  skip_v = "batch|lymphoid|myeloid|neoplastic"
)
}
\arguments{
\item{names_lsv}{list of vectors or just a single vector}

\item{order_v}{vector of matching treatments in the correct order}

\item{skip_v}{regex expression of list elements that should be skipped b/c they're not treatments.}
}
\value{
same object, but ordered
}
\description{
Order a vector of treatments either standalone or as part of a list.
}
