% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{orig_lymphoidColors_v}
\alias{orig_lymphoidColors_v}
\title{Original Data Lymphoid Colors}
\format{
\subsection{\code{orig_lymphoidColors_v}}{

Named vector of length 13
\describe{
\item{values}{Hex Color}
\item{names}{Sub Population Name}
}
}
}
\usage{
orig_lymphoidColors_v
}
\description{
Colors used on UMAPs for b1+b2 lymphoid populations
from original Ana/Rossin analysis
}
\keyword{datasets}
