% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHits.R
\name{plotHits}
\alias{plotHits}
\title{Plot Hits}
\usage{
plotHits(obj, deg_lsdt, genes_v, pop_v, name_v)
}
\arguments{
\item{obj}{seurat object}

\item{deg_lsdt}{list of deg results tables}

\item{genes_v}{vector of genes to display}

\item{pop_v}{name of the population (e.g. "TAMs" or "b12 TAMs"). Not used for searching so can be anything.}

\item{name_v}{name of genes_v for the plot}
}
\value{
list of lists. First list level is the plot type (volcano, violin, feature), while the second level
comprises all of the comparisons provided in deg_lsdt
}
\description{
Make volcano, violin, and feature plots
}
\details{
The names of deg_lsdt should be in format "Treat1_Treat2", the content of each list element
is the specific DEG table associated with Treat1 vs Treat2 comparison.
obj should be subset to have the same cells that are tested in the Treat1 vs Treat2 comparison.
genes_v is a vector of one or more genes that will be displayed in the plots.
name_v is a name that describes the gene(s) in genes_v
}
