% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMultiTargetChord.R
\name{plotMultiTargetChord}
\alias{plotMultiTargetChord}
\title{Plot Asymmetric Multi-Target Chord Diagram}
\usage{
plotMultiTargetChord(
  circData_dt,
  senderPop_v,
  receiverPop_v = NULL,
  senderDataCol_v,
  receiverDataCol_v,
  popCol_v = "Pop",
  sectorColors_v = NULL,
  arcColorCol_v = "Treatment",
  arcColors_v = NULL,
  legendType_v = "both",
  axisTicks_v = F,
  tickScaler_v = 5,
  cex_v = 1,
  specialLabel_v = F,
  startDegree_v = 270
)
}
\arguments{
\item{circData_dt}{Cleaned data.table.}

\item{senderPop_v}{Population to act as the sender}

\item{receiverPop_v}{The populations to act as receiver. NULL (default) uses all.}

\item{senderDataCol_v}{column that has values for sender}

\item{receiverDataCol_v}{column that has values for receiver}

\item{popCol_v}{character vector indicating which column of circData_dt has sector info}

\item{sectorColors_v}{optional vector to color sectors}

\item{arcColorCol_v}{column indicating what to color connectors by (usually treatment)}

\item{arcColors_v}{Named vector of colors.}

\item{legendType_v}{one of 'both', 'sector', 'arc', or 'none' indicating what to include in legend}

\item{axisTicks_v}{logical indicating to draw ticks or not}

\item{tickScaler_v}{what value should the tick mark scale bar show?}

\item{cex_v}{size of labels}

\item{specialLabel_v}{not sure how to add two sector labels, so have to use this and run a second time for combo figure.}

\item{startDegree_v}{degree on circle that arcs begin from. Default is 270, but have to change depending on relative sizes of each arc}
}
\value{
print plot
}
\description{
Generates a chord diagram for a single sender population linking to all
receiver populations.
}
\details{
Run with specialLabel_v = F for standard output. Change to true to get the sender/receiver labels that go
outside the plot.
}
