% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Plot Volcano}
\usage{
plotVolcano(
  data_dt,
  splitVar_v = "assay",
  runNames_v,
  geneCol_v,
  lfc_v,
  pval_v,
  ident1_v,
  colorCol_v,
  title_v = NULL,
  verbose_v = F
)
}
\arguments{
\item{data_dt}{data.table of gene expression. Can be wrangled for volcano or not. Must have ident1_v and another ident.}

\item{splitVar_v}{name to call the variable that the data is split on. If provided. If null, no split.}

\item{runNames_v}{values that splitVar_v is split on. If provided.}

\item{geneCol_v}{name of column name containing genes. Should be Gene and passed from parent funtion}

\item{lfc_v}{log-fold change to include in plot as cut-off}

\item{pval_v}{adjusted p-value to include in plot as cut-off}

\item{colorCol_v}{column to use for colors. Should be either diffExp for standard, or paste0(splitVar_v, "DE") for comparison ones.}

\item{title_v}{optional plot title.}

\item{verbose_v}{logical indicating to print messages.}
}
\value{
volcano plot
}
\description{
Make a volcano plot of DEG results
}
\details{
Make a volcano plot comparing the DEGs of two different groups. Can be one findmarker result, or can plot two sets of results using splitVar_v
}
