% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYScatter.R
\name{plotXYScatter}
\alias{plotXYScatter}
\title{Plot XY Scatter}
\usage{
plotXYScatter(data_dt, x_v, y_v, title_v = NULL, cor_v = T)
}
\arguments{
\item{data_dt}{data.table to plot. Must have columns corresponding to x_v and y_v that are numeric}

\item{x_v}{column name corresponding to x-axis variable}

\item{y_v}{column name corresponding to y-axis variable}

\item{title_v}{optional plot title}

\item{cor_v}{logical. Indicates whether or not to include regression line and correlation}
}
\value{
ggplot object
}
\description{
Plot XY scatter to compare two variables.
Include correlation, if desired.
}
\details{
Create a standard ggplot xy-scatterplot using the provided parameters.
No option to facet right now, best to pair with ggarrange()
Written to compare lfc and pvals of two different FindMarkers() runs, but can be used for any x-y pair.
}
