% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMultiTargetData.R
\name{prepareMultiTargetData}
\alias{prepareMultiTargetData}
\title{Prepare multi-target chord data}
\usage{
prepareMultiTargetData(
  circData_dt,
  senderPop_v,
  receiverPop_v = NULL,
  senderDataCol_v,
  receiverDataCol_v,
  popCol_v = "Pop",
  arcColorCol_v = "Treatment"
)
}
\arguments{
\item{circData_dt}{Cleaned data.table from the input CSV.}

\item{senderPop_v}{The population to act as the sender.}

\item{receiverPop_v}{The populations to act as receiver. NULL (default) uses all.}

\item{senderDataCol_v}{column that has values for sender}

\item{receiverDataCol_v}{column that has values for receiver}

\item{popCol_v}{character vector indicating which column of circData_dt has sector info}

\item{arcColorCol_v}{column indicating what to color arcs by (usually treatment)}
}
\value{
A data.table with \link{sender, receiver, treatment, senderWidth, receiverWidth}
}
\description{
Creates a link-level data.table where a single sender population is mapped
to all potential receiver populations for every treatment.
}
