% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepProjectR.R
\name{prepProjectR}
\alias{prepProjectR}
\title{Prep data for ProjectR Run}
\usage{
prepProjectR(
  obj,
  assay_v = "SCT",
  slot_v = "data",
  filter_dt = NULL,
  filterCol_v = "gene_name"
)
}
\arguments{
\item{obj}{seurat object}

\item{assay_v}{name of assay to get data from}

\item{filter_dt}{optional table containing at least one column of gene names that will be checked against genes in obj}

\item{filterCol_v}{column name from filter_dt used to identify genes.}
}
\value{
a data.frame of counts, filtered for provided genes.
}
\description{
Exctract correct assay/slot counts and filter
}
