% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepProjectRPlotting.R
\name{prepProjectRPlotting}
\alias{prepProjectRPlotting}
\title{Prep ProjectR Plotting}
\usage{
prepProjectRPlotting(
  sig_lssig,
  obj,
  slot_v = "data",
  pvals_v = c(0.01, 0.05),
  sigName_v = "sncScore",
  binaryName_v = "SnC"
)
}
\arguments{
\item{sig_lssig}{projection list object output by projectR (contains 'pval' and 'projection', 1-row arrays of same number of columns)}

\item{obj}{seurat object used to generate projections}

\item{slot_v}{slot used in projectR. Used for naming columns}

\item{sigName_v}{name to call the signature. Used for naming columns}

\item{binaryName_v}{character vector used to name cells that pass the filter. Cells will be 'binaryName_v' and 'not binaryName_v'}

\item{pval_v}{numeric vector of p-values to filter by}
}
\value{
updated seurat object with new columns.
ProjectR results:
\enumerate{
\item sig_lssig$pval will be an exact copy, called paste(slot_v, sigName_v, "pval", sep = "_")
\item sig_lssig$projection will be an exact copy, called paste(slot_v, sigName_v, "weight", sep = "_")
Wrangled ProjectR results (for each pvalue provided)
\item column called paste(slot_v, sigName_v, "weight_sig", pval_v, sep = "_") is a copy of the projection weights, but only for significant ones
\item column called paste(slot_v, pval_v, "sig", binaryName_v, sep = "_") binarizes the weights into "pos" and "neg" if they are significant
\item column called paste(slot_v, pval_v, binaryName_v, sep = "_") binarizes ALL the weights into "pos" and "neg"
For each pvalue provided in pval_v, two columns will
}
}
\description{
Add projection output to seurat object metadata
}
