% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenPathwayNames.R
\name{shortenPathwayNames}
\alias{shortenPathwayNames}
\title{Shorten Pathway Names}
\usage{
shortenPathwayNames(gsea_res)
}
\arguments{
\item{gsea_res}{input gsea results}
}
\value{
gsea_res, except 'pathway' column is shorter
}
\description{
shorten gsea pathway names to make output easier to read
}
