% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouetteScore.R
\name{silhouetteScore}
\alias{silhouetteScore}
\title{Score Silhouettes of Clusters}
\usage{
silhouetteScore(
  seurat_obj,
  resolution_v = 0.6,
  graph_v = "RNA_snn",
  ndims_v = 50,
  reduction_v = "pca",
  umap_v = "umap"
)
}
\arguments{
\item{seurat_obj}{A seurat object}

\item{resolution_v}{Resolution to use}

\item{graph_v}{name of graph to use.}

\item{ndims_v}{The number of dimensions used to run}

\item{reduction_v}{The reduction to use (pca, harmony, or iNMF, usually)}
}
\value{
Not sure yet
}
\description{
Given a seurat object and desired parameters, calculate clusters and determine silhouette scores for clusters.
}
