% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardPlots.R
\name{standardPlots}
\alias{standardPlots}
\title{Standard Seurat Plots}
\usage{
standardPlots(
  seurat_obj,
  reduction_v,
  clustCol_v,
  res_v,
  name_v,
  pt.size_v = 0.5,
  featurePlots_v = T,
  maxQ_v = "q90"
)
}
\arguments{
\item{seurat_obj}{seurat object to use for plotting}

\item{reduction_v}{name of reduction to use for UMAP - usually "umap" or "harmony"}

\item{clustCol_v}{column name of seurat clusters to plot}

\item{res_v}{resolution used for clusters}

\item{name_v}{name to add to plot titles}

\item{pt.size_v}{argument to pt.size of DimPlot and FeaturePlot}

\item{featurePlots_v}{logical indicating if set of feature plots should also be made.}

\item{maxQ_v}{cut off for feature plots. Use this if the scale is such that you can't see differences.}
}
\value{
list of ggplot objects
}
\description{
Make some quick plots for unintegrated, harmony, or other objects
}
