% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeProjectRScores.R
\name{summarizeProjectRScores}
\alias{summarizeProjectRScores}
\title{Summarize ProjectR Scores}
\usage{
summarizeProjectRScores(plot_df, cols_v = NULL)
}
\arguments{
\item{plot_df}{plot table output gy getProjectRPlottingData}

\item{cols_v}{columns to summarize by}
}
\value{
summary table
}
\description{
Get a summary of scores
}
