% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeProjectRScores.R
\name{tabulateScores}
\alias{tabulateScores}
\title{Tabulate Signature Scores}
\usage{
tabulateScores(plot_df, weight_v, cols_v = NULL)
}
\arguments{
\item{plot_df}{data.frame containing plotting information}

\item{weight_v}{column name of weight that is getting summarized}

\item{cols_v}{other columns in plot_df that should be used to stratify the summary of weight_v}
}
\value{
table with counts of each group
}
\description{
Summarize input table into Pos/Neg/Zero by provided variables
}
