% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treatColors_v}
\alias{treatColors_v}
\title{Treatment Colors}
\format{
\subsection{\code{treatColors_v}}{

Named vector of length 9
\describe{
\item{values}{Hex Color}
\item{names}{Treatment Name}
}
}
}
\usage{
treatColors_v
}
\description{
Colors used on plots for treatments
}
\keyword{datasets}
