% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twoGroupVolcanoColors_v}
\alias{twoGroupVolcanoColors_v}
\title{Volcano Color Scale - Two groups}
\format{
\subsection{\code{twoGroupVolcanoColors_v}}{

named vector of colors of length 6
\describe{
\item{values}{grey, blue, red, darkgrey, darkblue, darkred}
\item{names}{NO_1, DOWN_1, UP_1, NO_2, DOWN_2, UP_2}
}
}
}
\usage{
twoGroupVolcanoColors_v
}
\description{
Color scale to use in comparison (two-group)
volcano plots
red - higher expression in g1
blue - lower expression in g1
grey - no change in expression in g1
darkred - higher expression in g2
darkblue - lower expression in g2
darkgrey - no change in expression in g2
}
\keyword{datasets}
