% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlnPlotCompareMeans.R
\name{vlnPlotCompareMeans}
\alias{vlnPlotCompareMeans}
\title{Violin Plot Compare Means}
\usage{
vlnPlotCompareMeans(
  data_df,
  name_v,
  indVar_v = "Treatment",
  measureVar_v,
  groupBy_v = NULL,
  method_v = "wilcox.test",
  padjMethod_v = "holm",
  displayP_v = "padj",
  popCol_v = "sPop",
  comp_dt = NULL,
  colors_v = NULL,
  nrow_v = 1
)
}
\arguments{
\item{data_df}{data.frame containing some combination of groups and measure variables.}

\item{name_v}{name to describe data in data_df. Will be in plot title.}

\item{indVar_v}{character vector of independent variable, see details. Treatment is default}

\item{measureVar_v}{column name of column in data_dt that contains the metric to be plotted}

\item{groupBy_v}{optional value to facet plots by}

\item{method_v}{passed to stat::compare_means 'method' argument}

\item{padjMethod_v}{passed to stat::compare_means 'p.adjust.method' argument}

\item{displayP_v}{vector. If "padj", will display adjusted p-value. if numeric, will round p-value to that many digits.}

\item{popCol_v}{either 'sPop' or 'collapsePop'}

\item{comp_dt}{data.table of comparisons to subset calculated stats for. If null, will do all. See details.}

\item{colors_v}{named vector of hex codes. Names must be values of indVar_v}

\item{nrow_v}{number of rows to coerce output into.}
}
\value{
violin plot ggplot
}
\description{
Make a violin plot of some metric and compare groups
}
\details{
indVar_v is used to build a formula with measureVar_v: as.formula(paste0(measureVar_v, " ~ ", indVar_v))
comp_dt has two columns: group1 and group2. The values must be valid values in data_dt[\link{indVar_v}]
if groupBy_v is null, just one plot, if not, make one plot per unique value in group by and arrange with ggarrange
}
