% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoWrangleMarkers.R
\name{volcanoWrangleMarkers}
\alias{volcanoWrangleMarkers}
\title{Wrangle Marker Data for Volcano Plot}
\usage{
volcanoWrangleMarkers(
  data_dt,
  geneCol_v = "Gene",
  lfcCol_v = "avg_log2FC",
  lfc_v = 0.25,
  pvalCol_v = "p_val_adj",
  pval_v = 0.05,
  labelAll_v = T,
  lfcLabelCutOff_v = T
)
}
\arguments{
\item{data_dt}{data.table output from FindAllMarkers}

\item{geneCol_v}{name of column containing gene names. Default is Gene}

\item{lfcCol_v}{name of column containing log fold change values. Default is avg_log2FC}

\item{lfc_v}{log fold change cut-off for labels}

\item{pvalCol_v}{name of column contianing adjusted pvalues. Default is p_val_adj}

\item{pval_v}{pvalue cut off for labels}

\item{labelAll_v}{logical. TRUE - label all significant genes. FALSE - label top 25 only}

\item{lfcLabelCutOff_v}{logical. TRUE - only label genes that are significant and above the lfc_v threshold.}
}
\value{
original data_dt with updated metadata info for plotting
}
\description{
Add information for volcano plot
}
