% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wesSample_scRNATreat_dt}
\alias{wesSample_scRNATreat_dt}
\title{WES Sample - Treatment Mapping}
\format{
\subsection{\code{wesSample_scRNATreat_dt}}{

data.table with 22 rows and 2 columns
\describe{
\item{Tumor_Sample_Barcode}{Sample ID in WES data}
\item{Treatment}{Treatment code}
}
}
}
\usage{
wesSample_scRNATreat_dt
}
\description{
data.table that maps sample IDs from Whole Exome
Sequencing to treatment labels
}
\keyword{datasets}
