% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleCollapsePops.R
\name{wrangleCollapsePops}
\alias{wrangleCollapsePops}
\title{Wrangle Collapse Populations}
\usage{
wrangleCollapsePops(obj, name_v, batch_v, col_v = "sPop", newCol_v = NULL)
}
\arguments{
\item{obj}{seurat object}

\item{name_v}{name of seurat object. Usually b12_lymphoid or similar. Used to get map_dt!}

\item{batch_v}{name of batch. Used if name_v doesn't have batch in the name}

\item{col_v}{what column of meta.data are we updating? sPop is only one so far.}

\item{newCol_v}{name of column to put collapsed populations in. Default NULL will replace values in col_v}
}
\value{
Either: (1) same exact seurat object, but col_v is now updated with collapsed populations
(2) same seurat object with one more metadata column containing the collapsed populations
}
\description{
Use population map tables to collapse original sPop names
in seurat object meta.data
}
