% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangleDEG.R
\name{wrangleDEG}
\alias{wrangleDEG}
\title{Wrangle DEG}
\usage{
wrangleDEG(
  deg_dt,
  subList_lsv = list(neoantigen = allNeoantigens_v, immunogenic = allImmunoNeoantigens_v,
    top65Immuno = sortAllTop65ImmunoNeoantigens_v, atac4x = atacNeoantigenList_v)
)
}
\arguments{
\item{deg_dt}{data.table converted from FindMarkers data.frme output}

\item{subList_lsv}{list of genes to subset data by}
}
\value{
list of results subset variously
}
\description{
Wrangle DEGs with neoantigen labels and subset
}
\details{
This is used in the neoplasticDEG_checkAntigen reports. the referenced antigens in subList_lsv are loaded there.
}
