% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrh.scRNA-package.R
\docType{package}
\name{wrh.scRNA-package}
\alias{wrh.scRNA}
\alias{wrh.scRNA-package}
\title{wrh.scRNA: scRNAseq Analysis Helper Functions}
\description{
Helper functions for some common operations in the Seurat scRNAseq workflow and loadable objects containg color codes and other project-specific information.
}
\author{
\strong{Maintainer}: Wes Horton \email{hortowe@ohsu.edu}

}
\keyword{internal}
