% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkAtlasAnnotations.R
\name{checkAtlasAnnotations}
\alias{checkAtlasAnnotations}
\title{Check Atlas Annotation}
\usage{
checkAtlasAnnotations(
  clusterMarkers_dt,
  popMap_dt,
  popMarkerChecks_v,
  topClusterGenes_v = 50,
  topVolcGenes_v = 15
)
}
\arguments{
\item{clusterMarkers_dt}{data.table containing FindAllMarkers results.}

\item{popMap_dt}{data.table mapping seurat clusters to cell types}

\item{popMarkerChecks_v}{named vector of genes associated with various populations}

\item{topClusterGenes_v}{numeric indicating the number of genes to use to search for the pop marker checks in. Starting with the highest LFC. Default is 50, NULL to use all.}

\item{topGenes_v}{numeric indicating how many of the top genes to label in the volcano (fed to plotVolcano() labelTop_v argumennt)}
}
\value{
list of volcano plots of hits and data.table with intersections
}
\description{
Check each seurat cluster and cell type identity in Eric's public annotations
against the genes we have.
}
