% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displaySummary.R
\name{displaySummary}
\alias{displaySummary}
\title{Display Summary Info}
\usage{
displaySummary(
  obj,
  subObj = NULL,
  name_v = NULL,
  subCol_v = "mPop",
  batch_v,
  summary_lsv = list(treat = "Treatment", pop = "mPop", combo = c("Treatment", "mPop")),
  outDir_v = NULL,
  print_v = T,
  cutOff_v = cutOff_v
)
}
\arguments{
\item{obj}{seurat object to summarize}

\item{subObj}{subset object. need to update getSummaryTables to work on subObj, so i don't need both of these}

\item{name_v}{population name. will remove "batch#", if present. See details.}

\item{subCol_v}{Column to use in conjunction with name_v to subset data to summarize.}

\item{batch_v}{name of batch}

\item{summary_lsv}{list of vectors of treatments to summarize. See details.}

\item{outDir_v}{optional output directory to save plots and data}

\item{print_v}{logical indicating to print tables and plots to console.}

\item{cutOff_v}{numeric value indicating minimum number of cells needed in each group.}
}
\value{
no object returned. Two optional outputs:
\enumerate{
\item save plots and data to file
and/or
\item print plots to console.
}
}
\description{
Display summary tables and umaps of provided object.
}
\details{
This function calls getSummaryTables(), formats those tables, and makes umaps
of seurat clusters and population identities.

summary_lsv determines how the data are summarized. It is a named list where
each element's name is user defined and should describe what the output is summarizing.
Refer to the defaults for help. The values of each element correspond to the column or
columns that will be subset from obj@meta.data in order to generate that summary.

The default will summarize Treatment alone, major population alone, and treat x pop
together. Right now can only handle 'treat', 'pop' and 'combo'
}
