% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterProps.R
\name{plotClusterProps}
\alias{plotClusterProps}
\title{Plot Cluster Proportions}
\usage{
plotClusterProps(
  seurat_obj,
  clusters_v = "seurat_clusters",
  ident_v = NULL,
  legendName_v,
  clustColors_v,
  identColors_v = treatColors_v,
  sortBySize_v = F
)
}
\arguments{
\item{seurat_obj}{\itemize{
\item seurat object with at least one "seurat_clusters" column
}}

\item{clusters_v}{\itemize{
\item the name of the "seurat_clusters" column if not default name.
}}

\item{ident_v}{\itemize{
\item the name of the metadata column to use as grouping variable
}}

\item{legendName_v}{name to put in legend to label ident_v}

\item{clustColors_v}{named vector of colors that correspond to clusters_v column}

\item{identColors_v}{named vector of colors that correspond to ident_v}

\item{sortBySize_v}{logical, will sort by cluster size rather than factor levels}
}
\value{
Returns patchwork of two ggplots of cluster data plots
}
\description{
Plot proportions of idents falling in clusters. Adapted from
https://www.singlecellcourse.org/scrna-seq-dataset-integration.html#seurat-v3-3-vs-5-10k-pbmc
}
