% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoWrangleMarkers.R
\name{volcanoWrangleMarkers}
\alias{volcanoWrangleMarkers}
\title{Wrangle Marker Data for Volcano Plot}
\usage{
volcanoWrangleMarkers(
  data_dt,
  geneCol_v = "Gene",
  lfcCol_v = "avg_log2FC",
  lfc_v = 0.25,
  pvalCol_v = "p_val_adj",
  pval_v = 0.05,
  labelGenes_v = NULL,
  labelAll_v = T,
  labelTop_v = 20,
  labelDir_v = "both"
)
}
\arguments{
\item{data_dt}{data.table output from FindMarkers}

\item{geneCol_v}{name of column containing gene names. Default is Gene}

\item{lfcCol_v}{name of column containing log fold change values. Default is avg_log2FC}

\item{lfc_v}{log fold change cut-off for labels}

\item{pvalCol_v}{name of column contianing adjusted pvalues. Default is p_val_adj}

\item{pval_v}{p-value cut off for labels}

\item{labelGenes_v}{character vector of gene names to label. See details.}

\item{labelAll_v}{logical. TRUE - label all significant genes. See details.}

\item{labelTop_v}{numeric. if labelAll_v == F, number of significant genes to label}
}
\value{
original data_dt with updated metadata info for plotting
}
\description{
Add information for volcano plot
}
\details{
Reformat the default FindMarkers() DEG output to contain information we want for volcano plots.
\enumerate{
\item Binarize differential expression by assigning all significant genes (i.e. p-value less than pval_v) to either "UP"
or "DOWN" depending of if lfcCol_v is greater than or less than lfc_v, respectively.
\item Assign labels to specific genes in the plot (3 options). Default is option b.
a. If labelGenes_v is set (i.e. not NULL), then these genes will be labeled, if present. This takes precedence
b. If labelAll_v == T (and is.null(labelGenes_v)), then label all genes that pass the pvalue and lfc threshold
c. If labelAll_v == F (and is.null(labelGenes_v)), then label labelTop_v genes. If labelTop_v == 0, no labels.
}
}
